% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_K.R
\name{determine_K}
\alias{determine_K}
\title{Determine number of groups using information criterion}
\usage{
determine_K(N, TT, y, X, y_raw, X_raw, lambda_seq, K_max, rho = 2/3 * (N
  * TT)^(-0.5), FUN = PLS.cvxr, beta0 = NULL, MaxIter = 500,
  tol = 1e-04)
}
\arguments{
\item{N}{individual dimension}

\item{TT}{time dimension}

\item{y}{y(TN * 1)}

\item{X}{X(TN * P)}

\item{y_raw}{y_raw(TN * 1) raw data without standardization}

\item{X_raw}{X_raw(TN * P) raw data without standardization}

\item{lambda_seq}{Candidate tuning variable}

\item{K_max}{Maximum number of groups}

\item{rho}{Tuning parameter in the IC}

\item{FUN}{choose which function to be used in PLS estimation}

\item{beta0}{N*p matrix. initial estimator}

\item{MaxIter}{Maximum # of iteration}

\item{tol}{convergence criterion}
}
\value{
A list contains optimal K and lambda
\item{lambda}{Optimal lambda}
\item{K}{Optimal K}
}
\description{
Determine number of groups using information criterion
}
