% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPJ_PLS.R
\name{SPJ_PLS}
\alias{SPJ_PLS}
\title{Half panel Jackknife
develop on Dhaene and Jochmans (2015)}
\usage{
SPJ_PLS(t, y, x)
}
\arguments{
\item{t}{# of time periods}

\item{y}{(n*t vector) response variable}

\item{x}{(nt * p matrix)independent variable}
}
\description{
Half panel Jackknife
develop on Dhaene and Jochmans (2015)
}
