% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_est.R
\name{PLS.cvxr}
\alias{PLS.cvxr}
\title{PLS estimation by the iterative algorithm via CVXR}
\usage{
PLS.cvxr(N, TT, y, X, K, lambda, beta0 = NULL, R = 500, tol = 1e-04,
  solver = "ECOS", post_est = TRUE, bias_corr = FALSE)
}
\arguments{
\item{N}{The dimension of cross-sectional units in the panel.}

\item{TT}{The time series dimension in the panel.}

\item{y}{Dependent variable. (TN * 1). T is the fast index.}

\item{X}{Independent variable. (TN * P). T is the fast index. P is the number of regressors.}

\item{K}{The number of groups.}

\item{lambda}{The tuning parameter.}

\item{beta0}{N*p matrix. The initial estimator for each i=1,...,N.}

\item{R}{Maximum number of iteration.}

\item{tol}{Tolerance level in the convergence criterion.}

\item{solver}{'ECOS' or 'MOSEK'}

\item{post_est}{A boolean: do post-lasso estimation or not.}

\item{bias_corr}{A boolean: do bias correction in the post-lasso estimation or not.}
}
\value{
A list contains estimated coeffcients and group struncture
\item{b.est}{N * p matrix containing estimated slope for each individual}
\item{a.out}{K * p matrix containing estimated slope for each group}
\item{group.est}{group_id for each individual}
\item{converge}{A boolean indicating whether convergence criteria is met }
}
\description{
PLS estimation by the iterative algorithm via CVXR
}
