% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS_est.R
\name{PLS.mosek}
\alias{PLS.mosek}
\title{PLS estimation by the iterative algorithm via Rmosek}
\usage{
PLS.mosek(N, TT, y, X, K, lambda, beta0 = NULL, R = 500, tol = 1e-04,
  post_est = TRUE, bias_corr = TRUE)
}
\arguments{
\item{N}{individual dimension}

\item{TT}{time dimension}

\item{y}{y(TN * 1)}

\item{X}{X(TN * P)}

\item{K}{number of groups}

\item{lambda}{Tuning variable}

\item{beta0}{N*p matrix. initial estimator}

\item{R}{Maximum # of iteration}

\item{tol}{convergence criterion}
}
\value{
A list contains estimated coeffcients and group struncture
\item{b.est}{N * p matrix containing estimated slope for each individual}
\item{a.out}{K * p matrix containing estimated slope for each group}
\item{group.est}{group_id for each individual}
\item{converge}{A boolean indicating whether convergence criteria is met }
}
\description{
PLS estimation by the iterative algorithm via Rmosek
}
